requirements.txt
ipykernel
python-dotenv
langchain
langchain-openai
beautifulsoup4
langchain_community
faiss-cpu
streamlit
langchain_groq #LCEL
langchain-google-genai # Google Gemini (LCEL)
fastapi #LCEL
uvicorn #LCEL
langserve #LCEL
sse_starlette #LCEL
